requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
